+++ src/main/java/net/tassia/pancake/util/version/Version.kt
@@ -1,4 +1,4 @@
-package net.tassia.pancake
+package net.tassia.pancake.util.version
 
 import java.util.regex.Pattern
